+++ libs/src/svkImageData.h
@@ -215,7 +215,7 @@ class svkImageData: public vtkImageData
         void               GetIndexFromID(int voxelID, int* indexX, int* indexY, int* indexZ);
         void               GetIndexFromID(int voxelID, int* index);
         void               GetIndexFromPosition(double posLPS[3], int* index);
-        virtual void       GetPositionFromIndex(int* index, float* posLPS);
+        virtual void       GetPositionFromIndex(int* index, double* posLPS);
         virtual void       GetSliceOrigin(
                                 int slice, 
                                 float* sliceOrigin,
